copy: avoid useless comparison for non-directory entries
The comparison of the device number is meant to not follow directories
across multiple filesystems, so the test is only relevant when the
entry is a directory.

This also fixes the test-copy test when /tmp is an overlayfs with a
kernel older than Linux 4.12 since overlayfs reported inconsistent
devices numbers. See https://bugs.debian.org/854400 for some details.