core: don't change dbus property Job.Type (without a signal)
It is technically valid to change the property without emitting a signal.
However it seems strange.  The type change is mainly an implementation
detail.  State was changed at the same time as Type, and it emitted a
signal.

Preserve the original job type and use it for the dbus interface.