capability: fix CapabilityBoundingSet=~ does not work correctly.
    CapabilityBoundingSet=~CAP_XXX_A
    CapabilityBoundingSet=~CAP_XXX_B

If you use CapabilityBoundingSet=~CAP_XXX as an even number as in the case above,
the result is (~CAP_XXX_A | ~CAP_XXX_B) in the code before the modification.
It means ~0, with is equivalent to the CAP_ALL value.

The expected value is (~CAP_XXX_A & ~CAP_XXX_B),
which is the same value as ~(CAP_XXX_A | CAP_XXX_B).

Also, it does not logically clear the following items.
    CapabilityBoundingSet=~