Show failures which occur during `systemctl stop`
Fix #6478.

Also failures during stop could be part of a stop+start job aka restart.
We show the same message for this case.  It is marked "Warning: ", to try
and show that it is not a fatal error, and we continued with the start.

If there are failures during both stop and start, the messages get a little
messy :(.  We add an extra message between the two failures.  Without it
the second failure message would appear to be extra details about the stop
failure.

# Conflicts:
#	test/TEST-03-JOBS/test-jobs.sh