nspawn: Add --syscall-filter option
Adds a --syscall-filter option to set custom system call filters.
After the comments in #5944 this now works similar to the `SystemCallFilter=`
option in unit files.

This option takes a comma-separated list of system calls to allow.
If the first character of the list is "~", the listed system calls
will be blocked instead. Filter sets beginning with '@' are also
supported. This overwrites nspawn default blacklist and the `--[drop-]capability`
options

Fixes #5163