thunderbolt: initial support
This enables the support for thunderbolt 3 devices on kernel 4.13+.
Thunderbolt 3 devices need to be authorized via sysfs before they
can be used. There are two different modes:
  "user":   enable the devices by writing "1" to <device>/authorized
  "secure": prepare a key-challenge of the devices by writing a
            previously stored key to <device>/key and trigger the
            process via the string "2" written to <device>/authorized
The key can be stored in the device by writing it to <device>/key
and then manually authorizing the device via "1" to <device>/authorized

The database with devices to authorize is located in /etc/thunderbolt
and the authorization information is store either there or of EFI is
supported in efivars. The latter enables the use of thunderbolt devices,
e.g. Docks with USB keyboards attached, during the boot process.

The idea is that the initial authorization of unknown devices would be
triggered by desktop environments, e.g. GNOME, via "tbtctl authorize <dev>".
After this initial authorization, the devices would then automatically be
authorized via a udev rule that fetches the authorization information from
the store.

The store layout is:
  /etc/thunderbolt/

          devices/<uuid>       ini style file with device info
                               e.g. name, vendor

          authorization/<uuid> a) file or a symlink to the corresponding efi
                                  variable containing:

                                  "0": ignore this device
                                  "1": authorize this devices
                                  key: authorize via a key