login: dynamically join 'users' if 'at-console'
This extends pam_systemd.so to support the 'auth' target, alongside the
current 'session' target. When used with 'auth', it hooks into the
pam_setcred() call and dynamically joins the process to the 'users'
group, if, and only if, the session is considered to be 'at-console' (in
the D-Bus sense).

The idea here is to get an authoritative credential for every process
to be identified as 'at-console'. With the new SO_PEERGROUPS
socket-option, this will allow system daemons to authenticate their
peers through an AF_UNIX socket. Moreover, this will allow us to finally
get rid of the 'at_console="{true,false}"' policies in dbus-daemon(1),
and instead turn them into 'group="users"' policies.