logind: fix `loginctl enable-linger`
`man systemd-run`: "In the default configuration, users can enable
lingering for themselves".

But when I sshed in and tried it, I was prompted for the password of the
first user with `sudo` access (`wheel` group) or `root`.  Even though we
seemed to explicitly allow all uses in #3373.

After some code analysis: `XDG_SESSION_ID= loginctl enable-linger` worked
fine.

I don't know exactly why loginctl uses XDG_SESSION_ID.  But it is more
friendly anyway, if logind allows it when the value for UID is passed
explicitly, given that it would allow if it detected the same UID
implicitly after being passed UID_INVALID.