hwdb: add Chalkboard Electronics HID Touchscreen support
The Chalkboard Electronics HID Touchscreen is classified as
a tablet device because it has BTN_TOOL_PEN support. It's
actually a capacitive touchscreen IPS display panel and
should be reported as such. It also reports a resolution
of 0 for both axes in absinfo.

This hwdb entry fixes these issues by reclassifying the
device as a touchscreen and adding valid resolution values
in absinfo.

Signed-off-by: Matt Porter <mporter@konsulko.com>