networkd: ndisc client do not set IPv6 address lifetime 0.
Since we wre setting the lifetime valid and lifetime preferred
even if they are 0 it shows as deprecated.

in link_check_ready we look for whether address is depricated and
since this is depricated we never configure this like.

Also The preferred lifetime is never greater than the valid lifetime.

Fixes #6359